Safari-611.1.10/LayoutTests/fast/selectors/tagname-descendant-backtracking-case-sensitivity-html.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
    <container style="display:none">
        <container class="target">
            <container>
                <container>
                    <container>
                        <padding>
                            <target></target>
                        </padding>
                    </container>
                </container>
            </container>
        </container>
    </container>
</body>
<script>
description("Test the correctness of the backtracking optimizations with mixed case tag names in the selector.");

shouldBe('document.querySelectorAll("CONTAINER.target > CONTAINER > container target").length', '1');
</script>
<script src="../../resources/js-test-post.js"></script>
</html>
